+++ enterprise/frontend/src/embedding-sdk-bundle/components/public/CreateDashboardModal/CreateDashboardModal.tsx
@@ -3,7 +3,10 @@ import {
   getCollectionIdSlugFromReference,
   getCollectionIdValueFromReference,
 } from "embedding-sdk-bundle/store/collections";
-import type { MetabaseDashboard, SdkCollectionId } from "embedding-sdk-package";
+import type {
+  MetabaseDashboard,
+  SdkCollectionId,
+} from "embedding-sdk-bundle/types";
 import { useCollectionQuery, useLocale } from "metabase/common/hooks";
 import { CreateDashboardModal as CreateDashboardModalCore } from "metabase/dashboard/containers/CreateDashboardModal";
 import { useSelector } from "metabase/lib/redux";
