+++ enterprise/frontend/src/embedding-sdk-bundle/store/auth/index.ts
@@ -1,2 +1,1 @@
 export * from "./auth";
-export * from "../../../embedding/auth-common/saml-token-storage";
