+++ enterprise/frontend/src/embedding-sdk-bundle/store/reducer.ts
@@ -1,27 +1,16 @@
-import {
-  type AsyncThunkAction,
-  createAction,
-  createReducer,
-} from "@reduxjs/toolkit";
+import { createAction, createReducer } from "@reduxjs/toolkit";
 
-import { samlTokenStorage } from "embedding/auth-common";
-import type { SdkState, SdkStoreState } from "embedding-sdk-bundle/store/types";
-import type { MetabaseAuthConfig } from "embedding-sdk-bundle/types/auth-config";
+import type { SdkState } from "embedding-sdk-bundle/store/types";
 import type { SdkEventHandlersConfig } from "embedding-sdk-bundle/types/events";
 import type { MetabasePluginsConfig } from "embedding-sdk-bundle/types/plugins";
-import type {
-  MetabaseEmbeddingSessionToken,
-  MetabaseFetchRequestTokenFn,
-} from "embedding-sdk-bundle/types/refresh-token";
 import type {
   SdkErrorComponent,
   SdkLoadingError,
 } from "embedding-sdk-bundle/types/ui";
 import type { SdkUsageProblem } from "embedding-sdk-bundle/types/usage-problem";
-import { createAsyncThunk } from "metabase/lib/redux";
+import type { MetabaseFetchRequestTokenFn } from "metabase/embedding-sdk/types/refresh-token";
 
 import { initAuth, refreshTokenAsync } from "./auth";
-import { getSessionTokenState } from "./selectors";
 const SET_METABASE_INSTANCE_VERSION = "sdk/SET_METABASE_INSTANCE_VERSION";
 const SET_METABASE_CLIENT_URL = "sdk/SET_METABASE_CLIENT_URL";
 const SET_LOADER_COMPONENT = "sdk/SET_LOADER_COMPONENT";
@@ -44,52 +33,6 @@ export const setErrorComponent = createAction<null | SdkErrorComponent>(
 export const setError = createAction<SdkLoadingError | null>(SET_ERROR);
 export const setFetchRefreshTokenFn =
   createAction<null | MetabaseFetchRequestTokenFn>(SET_FETCH_REQUEST_TOKEN_FN);
-
-const GET_OR_REFRESH_SESSION = "sdk/token/GET_OR_REFRESH_SESSION";
-
-let refreshTokenPromise: ReturnType<
-  AsyncThunkAction<MetabaseEmbeddingSessionToken | null, unknown, any>
-> | null = null;
-
-export const getOrRefreshSession = createAsyncThunk(
-  GET_OR_REFRESH_SESSION,
-  async (
-    authConfig: Pick<
-      MetabaseAuthConfig,
-      "metabaseInstanceUrl" | "preferredAuthMethod"
-    >,
-    { dispatch, getState },
-  ) => {
-    // necessary to ensure that we don't use a popup every time the user
-    // refreshes the page
-    const storedAuthToken = samlTokenStorage.get();
-    const state = getSessionTokenState(getState() as SdkStoreState);
-    /**
-     * @see {@link https://github.com/metabase/metabase/pull/64238#discussion_r2394229266}
-     *
-     * TODO: I think this should be called session overall e.g. state.session
-     */
-    const session = storedAuthToken ?? state?.token;
-
-    const shouldRefreshToken =
-      !session ||
-      (typeof session?.exp === "number" && session.exp * 1000 < Date.now());
-    if (!shouldRefreshToken) {
-      return session;
-    }
-
-    if (refreshTokenPromise) {
-      return refreshTokenPromise.unwrap();
-    }
-
-    refreshTokenPromise = dispatch(refreshTokenAsync(authConfig));
-    refreshTokenPromise.finally(() => {
-      refreshTokenPromise = null;
-    });
-
-    return refreshTokenPromise.unwrap();
-  },
-);
 
 const SET_PLUGINS = "sdk/SET_PLUGINS";
 export const setPlugins = createAction<MetabasePluginsConfig | null>(
