+++ enterprise/frontend/src/embedding-sdk-bundle/test/sdk-config-errors/sdk-config-errors.unit.spec.tsx
@@ -1,3 +1,6 @@
+// Activates the plugins needed for the tests before any SDK code
+import "metabase-enterprise/sdk-plugins";
+
 import { render, screen, waitFor } from "@testing-library/react";
 
 import {
