+++ enterprise/frontend/src/embedding-sdk-ee/auth/index.ts
@@ -1,0 +1,2 @@
+// side effect inside `./auth.ts`
+import "./auth";
