+++ enterprise/frontend/src/embedding-sdk-ee/metabot/MetabotChatSuggestions.tsx
