+++ enterprise/frontend/src/embedding-sdk-ee/metabot/MetabotQuestion.module.css
