+++ enterprise/frontend/src/embedding-sdk-package/index.ts
@@ -122,13 +122,12 @@ export type {
   SqlParameterValues,
 } from "embedding-sdk-bundle/types";
 
-export type {
-  UserBackendJwtResponse,
-  MetabaseFetchRequestTokenFn,
-  MetabaseEmbeddingSessionToken,
-} from "embedding-sdk-bundle/types/refresh-token";
-
 export type { EmbeddingEntityType } from "metabase-types/store/embedding-data-picker";
 
 export type { ParameterValues } from "metabase/embedding-sdk/types/dashboard";
 export type { IconName } from "metabase/embedding-sdk/types/icon";
+
+export type {
+  UserBackendJwtResponse,
+  MetabaseFetchRequestTokenFn,
+} from "metabase/embedding-sdk/types/refresh-token";
