+++ enterprise/frontend/src/embedding/auth-common/index.ts
@@ -2,4 +2,3 @@ export { connectToInstanceAuthSso } from "./connect-to-instance-auth-sso";
 export { jwtDefaultRefreshTokenFunction } from "./jwt";
 export { openSamlLoginPopup } from "./saml";
 export { validateSession } from "./validate-session-token";
-export { samlTokenStorage } from "./saml-token-storage";
