+++ enterprise/frontend/src/metabase-enterprise/sdk-plugins.js
@@ -14,3 +14,8 @@ import "./embedding-sdk";
 import "./metabot";
 
 initializeSdkPlugins();
+
+// "SDK EE-plugins", that are specific to the embedding sdk.
+// These only apply to the SDK, not to the core app
+import "../embedding-sdk-ee/auth";
+import "../embedding-sdk-ee/metabot";
